// src/lib/types.ts
export type SemanticShorthandItemFunc<TProps> = (component: React.ReactType<TProps>, props: TProps, children?: React.ReactNode | React.ReactNodeArray) => React.ReactElement<any> | null;
export type SemanticShorthandContent = React.ReactNode;
export type SemanticShorthandItem<TProps> = React.ReactNode | TProps | SemanticShorthandItemFunc<TProps>;